refactor todo example
