+++ examples/flutter_main/lib/todo/models.dart
@@ -1,0 +1,21 @@
+part of 'todo.dart';
+
+enum Filter { all, active, done }
+
+@immutable
+class Todo {
+  const Todo({
+    required this.description,
+    required this.id,
+    this.completed = false,
+  });
+
+  final String id;
+  final String description;
+  final bool completed;
+
+  @override
+  String toString() {
+    return 'Todo(description: $description, completed: $completed)';
+  }
+}
