+++ examples/flutter_main/lib/todo/todo.dart
@@ -1,0 +1,10 @@
+import 'package:example/const.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:state_beacon/state_beacon.dart';
+
+import 'dart:math' as math;
+
+part 'controller.dart';
+part 'view.dart';
+part 'models.dart';
